+++ lib/pos_printer_bloc.dart
@@ -4,3 +4,4 @@ export 'src/bloc/printer_bloc.dart';
 export 'src/bloc/printer_scan_bloc.dart';
 export 'src/pos_ticket_line.dart';
 export 'src/printer.dart';
+export 'src/printer_scanner_builder.dart';