+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2] - 18 June 2020
+
+* Added basic Printer Scanner Builder
+* There's an issue with Star printers scanning on IOS - whole device is frozen during port scan
+
+## [0.0.1] - 18 June 2020
 
 Very early development version.
 
 * All errors handling is relying on onError of bloc's delegate
+
+